+++ Podcast/SeriesDetailViewController.swift
@@ -77,6 +77,7 @@ class SeriesDetailViewController: ViewController, SeriesDetailHeaderViewDelegate
         }
         
         loadingAnimation.startAnimating()
+        
     }
     
     override func viewDidAppear(_ animated: Bool) {
